@@ -1,28 +1,27 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
-import { RouterModule } from '@angular/router';
-import { Ng2SmartTableModule } from 'ng2-smart-table';
+import { BrowserModule } from "@angular/platform-browser";
+import { NgModule } from "@angular/core";
+import { FormsModule } from "@angular/forms";
+import {
+  provideHttpClient,
+  withInterceptorsFromDi,
+} from "@angular/common/http";
+import { RouterModule } from "@angular/router";
+import { Ng2SmartTableModule } from "ng2-smart-table";
 
-import { PagesModule } from './pages/pages.module';
-import { AppComponent } from './app.component';
-import { routes } from './app.routes';
-import { ScrollPositionDirective } from './theme/directives/scrollPosition.directive';
+import { PagesModule } from "./pages/pages.module";
+import { AppComponent } from "./app.component";
+import { routes } from "./app.routes";
+import { ScrollPositionDirective } from "./theme/directives/scrollPosition.directive";
 @NgModule({
-  declarations: [
-    AppComponent,
-    ScrollPositionDirective,
-  ],
+  declarations: [AppComponent, ScrollPositionDirective],
+  bootstrap: [AppComponent],
   imports: [
     BrowserModule,
     FormsModule,
-    HttpClientModule,
     RouterModule.forRoot(routes, { useHash: true }),
     Ng2SmartTableModule,
     PagesModule,
   ],
-  providers: [],
-  bootstrap: [AppComponent],
+  providers: [provideHttpClient(withInterceptorsFromDi())],
 })
-export class AppModule { }
+export class AppModule {}
